refactor: move deadline bar to ui components
