+++ project/app/(dashboard)/calendar/page.tsx
@@ -2,8 +2,8 @@
 
 import Button from "@/components/button";
 import Calendar from "@/components/calendar";
-import DeadlineBar from "@/components/deadline-bar";
 import { DashboardContent } from "@/components/layouts/dashboard/dashboard-content";
+import DeadlineBar from "@/components/ui/deadline-bar";
 import { Plus } from "lucide-react";
 
 export default function CalendarPage() {
